scripts/requirements.txt
Pillow
boto3
numpy
pandas
pyarrow
seaborn
ipykernel
mysqlclient
openai
retry
nanoid